requirements.txt
discord.py>=2.0.0,<3.0.0
asyncpg
jishaku==2.5.1